requirements.txt
imageio
pillow
# tensorflow
tf-nightly
loguru
fire
scikit-image
tensorflow_addons
tensorflow-model-optimization